% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.test.R
\name{cv.test}
\alias{cv.test}
\title{Cramer's V}
\usage{
cv.test(x, y)
}
\arguments{
\item{x}{a vector (categorical or numerical values)}

\item{y}{a vector (categorical or numerical values)}
}
\value{
Cramer's V
}
\description{
\code{cv.test} returns the Cramer's V correlation coefficient
}
\details{
The function calculates Cramer's V based on the results
of an Chi-Square-Test of Independence between two
categorical variables
}
\examples{
cv.test(x = iris$Species, iris$Sepal.Length)
}
