% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureCorMatrix.R
\name{plotFeatureCorMatrix}
\alias{plotFeatureCorMatrix}
\title{Title}
\usage{
plotFeatureCorMatrix(listCorMatrix)
}
\arguments{
\item{listCorMatrix}{list with matrix values and matrix type}
}
\value{
heatmap plot corresponding to the matrix values
}
\description{
Title
}
\examples{

}
