% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculates the Intraclass correlation}
\usage{
icc(depvar, indvar)
}
\arguments{
\item{depvar}{dependent variable, must be numeric}

\item{indvar}{independent variable, must be categorical}
}
\value{
returns the Intraclass correlation
}
\description{
The function calculates the Intraclass correlation
based on the results of the `aov` function
}
\examples{
icc(depvar = iris$Sepal.Length, indvar = iris$Species)
}
