% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intraclass Correlation (ICC)}
\usage{
icc(depvar, indvar)
}
\description{
\code{icc} returns the intraclass correlation coefficient
}
\details{
The function calculates the Intraclass Correlation Coefficient
based on the results of the aov() function
}
